Abstract section logic into a base component which can be inherited by all section components
