+++ src/web/app/pages-help/instructor-help-page/instructor-help-section/instructor-help-section.component.html
@@ -1,0 +1,3 @@
+<p>
+  instructor-help-section works!
+</p>
